split kv_router so it stops hoarding
