+++ lib/llm/src/kv_router/publisher/tests.rs
@@ -1,1049 +1,17 @@
-// SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-use std::future::Future;
-use std::sync::Arc;
-use std::sync::atomic::{AtomicU32, AtomicU64, Ordering};
-use std::time::{Duration, Instant};
-
-use anyhow::Result;
-use rmp_serde as rmps;
-use tokio::sync::mpsc;
-use tokio_util::sync::CancellationToken;
-use zeromq::{Socket, SocketRecv, SubSocket};
-
-use dynamo_runtime::metrics::MetricsHierarchy;
-use dynamo_runtime::traits::DistributedRuntimeProvider;
-use dynamo_runtime::transports::event_plane::EventPublisher;
-use dynamo_runtime::{
-    component::{Component, Namespace},
-    transports::nats::{NatsQueue, Slug},
-};
-
-/// Helper function to create a KV stream name from a component and subject.
-///
-/// Generates a slugified stream name in the format:
-/// `namespace-{namespace}-component-{component}-{subject}`
-fn create_kv_stream_name(component: &Component, subject: &str) -> String {
-    Slug::slugify(&format!(
-        "namespace.{}.component.{}.{}",
-        component.namespace().name(),
-        component.name(),
-        subject
-    ))
-    .to_string()
-    .replace("_", "-")
-}
-
-use dynamo_kv_router::indexer::{KvIndexerMetrics, LocalKvIndexer};
-use dynamo_kv_router::protocols::*;
-pub use dynamo_kv_router::zmq_wire::create_stored_blocks;
-use dynamo_kv_router::zmq_wire::*;
-
-use crate::kv_router::{
-    KV_EVENT_SUBJECT, KV_METRICS_SUBJECT, WORKER_KV_INDEXER_BUFFER_SIZE,
-    worker_query::start_worker_kv_query_endpoint,
-};
-use dynamo_runtime::config::environment_names::nats as env_nats;
-
-// Error handling configuration for ZMQ operations
-const INITIAL_BACKOFF_MS: u64 = 10;
-const MAX_BACKOFF_MS: u64 = 5000;
-const MAX_CONSECUTIVE_ERRORS: u32 = 10;
-const MAX_BACKOFF_EXPONENT: u32 = 8; // Cap at 2^8 = 256x multiplier to prevent overflow
-
-// Batching configuration
-const MAX_BATCHING_TIMEOUT_MS: u64 = 15_000; // 15 seconds, prevents misconfiguration
-pub const DEFAULT_BATCHING_TIMEOUT_MS: Option<u64> = None; // disabled by default
-const DEFAULT_MAX_BATCH_BLOCKS: usize = 128; // Max blocks to batch before flushing
-
-// ---------------------------------------------------------------------------
-// Engines dropped events metric
-// ---------------------------------------------------------------------------
-
-use std::sync::OnceLock;
-
-use dynamo_runtime::metrics::prometheus_names::kv_publisher;
-
-/// Metrics for the KV publisher, created via the MetricsHierarchy API.
-/// This provides automatic `dynamo_namespace`, `dynamo_component`, and other
-/// hierarchy labels for free.
-pub struct KvPublisherMetrics {
-    /// Total number of raw events dropped by engines before reaching publisher
-    pub engines_dropped_events_total: prometheus::IntCounterVec,
-}
-
-static KV_PUBLISHER_METRICS: OnceLock<Arc<KvPublisherMetrics>> = OnceLock::new();
-
-impl KvPublisherMetrics {
-    /// Create from a Component, memoized in a static OnceLock.
-    /// Uses the MetricsHierarchy API which auto-prepends `dynamo_component_`,
-    /// injects hierarchy labels, and registers with the DRT `MetricsRegistry`.
-    pub fn from_component(component: &Component) -> Arc<Self> {
-        KV_PUBLISHER_METRICS
-            .get_or_init(|| {
-                let metrics = component.metrics();
-                match metrics.create_intcountervec(
-                    kv_publisher::ENGINES_DROPPED_EVENTS_TOTAL,
-                    "Total number of raw events dropped by engines before reaching publisher (detected via event_id gaps)",
-                    &["worker_id"],
-                    &[],
-                ) {
-                    Ok(engines_dropped_events_total) => {
-                        Arc::new(Self { engines_dropped_events_total })
-                    }
-                    Err(e) => {
-                        tracing::warn!("Failed to create kv_publisher metrics from component: {}. Using unregistered metrics as fallback.", e);
-                        Arc::new(Self::new_unregistered())
-                    }
-                }
-            })
-            .clone()
-    }
-
-    /// Creates unregistered metrics for use when the MetricsRegistry is not available.
-    /// This is used as a fallback when metric creation fails.
-    pub fn new_unregistered() -> Self {
-        Self {
-            engines_dropped_events_total: prometheus::IntCounterVec::new(
-                prometheus::Opts::new(
-                    kv_publisher::ENGINES_DROPPED_EVENTS_TOTAL,
-                    "Total number of raw events dropped by engines before reaching publisher (detected via event_id gaps)",
-                ),
-                &["worker_id"],
-            )
-            .expect("failed to create engines_dropped_events_total counter"),
-        }
-    }
-
-    /// Increment the engines dropped events counter by the given amount.
-    pub fn increment_engines_dropped_events(&self, worker_id: u64, count: u64) {
-        self.engines_dropped_events_total
-            .with_label_values(&[&worker_id.to_string()])
-            .inc_by(count);
-    }
-}
-
-/// Get the KV publisher metrics if initialized.
-fn kv_publisher_metrics() -> Option<Arc<KvPublisherMetrics>> {
-    KV_PUBLISHER_METRICS.get().cloned()
-}
-
-// -------------------------------------------------------------------------
-// Batching State -----------------------------------------------------------
-// -------------------------------------------------------------------------
-
-/// Accumulator for in-flight KV cache events that will be merged into a single
-/// [`RouterEvent`] before being forwarded to the event sink.
-#[derive(Debug)]
-struct BatchingState {
-    /// Block hashes accumulating for the next Removed event.
-    pending_removed: Option<KvCacheRemoveData>,
-    /// Blocks accumulating for the next Stored event.
-    pending_stored: Option<KvCacheStoreData>,
-    /// Monotonic published-batch counter. Increments by 1 per flush so downstream
-    /// consumers always see consecutive event IDs, regardless of how many raw source
-    /// events were merged into the batch.
-    next_publish_id: u64,
-    /// dp_rank of the events in the current pending batch.
-    /// A change signals that the batch must be flushed before accumulating further.
-    last_dp_rank: u32,
-    /// When we last flushed (or initialized). Used to detect stale pending data:
-    /// if a new event arrives after a long idle period (exceeding timeout),
-    /// we flush immediately for lower latency on sparse important events.
-    last_flush_time: Instant,
-}
-
-impl BatchingState {
-    fn new() -> Self {
-        Self {
-            pending_removed: None,
-            pending_stored: None,
-            next_publish_id: 1,
-            last_dp_rank: 0,
-            last_flush_time: Instant::now(),
-        }
-    }
-
-    fn has_pending(&self) -> bool {
-        self.pending_removed.is_some() || self.pending_stored.is_some()
-    }
-
-    fn pending_block_count(&self) -> usize {
-        self.pending_removed
-            .as_ref()
-            .map(|r| r.block_hashes.len())
-            .unwrap_or(0)
-            + self
-                .pending_stored
-                .as_ref()
-                .map(|s| s.blocks.len())
-                .unwrap_or(0)
-    }
-
-    /// Records that a flush just happened. Called after every flush to track
-    /// idle periods for stale-data detection.
-    fn record_flush_time(&mut self) {
-        self.last_flush_time = Instant::now();
-    }
-
-    /// Returns the time remaining in the current batch window (zero if already elapsed).
-    fn remaining_timeout(&self, timeout_ms: u64) -> Duration {
-        let timeout = Duration::from_millis(timeout_ms);
-        let elapsed = self.last_flush_time.elapsed();
-        if elapsed >= timeout {
-            Duration::ZERO
-        } else {
-            timeout - elapsed
-        }
-    }
-
-    /// Returns `true` when the batch window has elapsed (or `timeout_ms` is zero).
-    fn is_timeout_elapsed(&self, timeout_ms: u64) -> bool {
-        self.remaining_timeout(timeout_ms) == Duration::ZERO
-    }
-}
-
-// -------------------------------------------------------------------------
-// KV Event Publishers -----------------------------------------------------
-// -------------------------------------------------------------------------
-
-/// Configure the source of KV events.
-/// Currently, only ZMQ is supported.
-pub enum KvEventSourceConfig {
-    Zmq { endpoint: String, topic: String },
-}
-
-/// The source of KV events.
-enum KvEventSource {
-    Zmq {
-        zmq_handle: tokio::task::JoinHandle<()>,
-    },
-}
-
-impl KvEventSource {
-    /// Start the event source from a [`KvEventSourceConfig`].
-    fn start(
-        component: Component,
-        worker_id: WorkerId,
-        kv_block_size: u32,
-        source_config: KvEventSourceConfig,
-        cancellation_token: CancellationToken,
-        tx: mpsc::UnboundedSender<PlacementEvent>,
-        next_event_id: Arc<AtomicU64>,
-    ) -> Result<Self> {
-        match source_config {
-            KvEventSourceConfig::Zmq { endpoint, topic } => {
-                let zmq_handle = component
-                    .drt()
-                    .runtime()
-                    .secondary()
-                    .spawn(start_zmq_listener(
-                        endpoint,
-                        topic,
-                        worker_id,
-                        tx,
-                        cancellation_token.clone(),
-                        kv_block_size,
-                        next_event_id,
-                    ));
-
-                Ok(KvEventSource::Zmq { zmq_handle })
-            }
-        }
-    }
-
-    fn shutdown(&self) {
-        match self {
-            KvEventSource::Zmq { zmq_handle } => {
-                zmq_handle.abort();
-            }
-        }
-    }
-}
-
-/// A publisher of KV events.
-pub struct KvEventPublisher {
-    /// The size of the KV block.
-    kv_block_size: u32,
-    /// The source of KV events.
-    /// Can be `None` if all events provided through [`KvEventPublisher::publish`].
-    source: Option<KvEventSource>,
-    /// The cancellation token.
-    cancellation_token: CancellationToken,
-    /// The ID of the local worker emitting placement events.
-    worker_id: WorkerId,
-    /// The channel to send events to.
-    tx: mpsc::UnboundedSender<PlacementEvent>,
-    /// Internal monotonic event ID counter - ensures each event gets a unique, incrementing ID.
-    /// Shared with the ZMQ listener (if any) to maintain consistency.
-    next_event_id: Arc<AtomicU64>,
-}
-
-impl KvEventPublisher {
-    pub fn new(
-        component: Component,
-        kv_block_size: u32,
-        source_config: Option<KvEventSourceConfig>,
-    ) -> Result<Self> {
-        Self::new_with_local_indexer(
-            component,
-            kv_block_size,
-            source_config,
-            false,
-            0,
-            DEFAULT_BATCHING_TIMEOUT_MS,
-        )
-    }
-
-    pub fn new_with_local_indexer(
-        component: Component,
-        kv_block_size: u32,
-        source_config: Option<KvEventSourceConfig>,
-        enable_local_indexer: bool,
-        dp_rank: DpRank,
-        batching_timeout_ms: Option<u64>,
-    ) -> Result<Self> {
-        let cancellation_token = CancellationToken::new();
-        // None = disabled (flush every event); Some(0) normalised to None; Some(ms) = opt-in.
-        // Cap at MAX_BATCHING_TIMEOUT_MS to prevent misconfiguration.
-        let batching_timeout_ms = batching_timeout_ms
-            .filter(|&ms| {
-                if ms > MAX_BATCHING_TIMEOUT_MS {
-                    tracing::warn!(
-                        requested_ms = ms,
-                        max_ms = MAX_BATCHING_TIMEOUT_MS,
-                        "batching_timeout_ms too high, capping to 15s"
-                    );
-                }
-                // if ms is 0, treat as disabled (None)
-                ms > 0
-            })
-            .map(|ms| ms.min(MAX_BATCHING_TIMEOUT_MS));
-
-        let (tx, rx) = mpsc::unbounded_channel::<PlacementEvent>();
-
-        // Infer worker_id from component's connection
-        let worker_id = component.drt().connection_id();
-
-        // Initialize the KV publisher metrics via MetricsHierarchy API
-        // This provides automatic hierarchy labels (dynamo_namespace, dynamo_component, etc.)
-        KvPublisherMetrics::from_component(&component);
-
-        let component_name = component.name();
-        tracing::info!(
-            "Initializing KvEventPublisher for worker {worker_id} in component {component_name}"
-        );
-
-        if enable_local_indexer {
-            tracing::info!(
-                "LocalKvIndexer enabled for worker {worker_id} in component {component_name}"
-            );
-        }
-
-        // Internal monotonic event ID counter - shared with ZMQ listener if any
-        let next_event_id = Arc::new(AtomicU64::new(0));
-
-        // Create our event source (if any)
-        let mut source = None;
-        if let Some(config) = source_config {
-            source = Some(KvEventSource::start(
-                component.clone(),
-                worker_id,
-                kv_block_size,
-                config,
-                cancellation_token.clone(),
-                tx.clone(),
-                next_event_id.clone(),
-            )?);
-        }
-
-        // Create local indexer if requested
-        let local_indexer = if enable_local_indexer {
-            let metrics = Arc::new(KvIndexerMetrics::new_unregistered());
-            Some(Arc::new(LocalKvIndexer::new(
-                cancellation_token.clone(),
-                kv_block_size,
-                metrics,
-                WORKER_KV_INDEXER_BUFFER_SIZE,
-            )))
-        } else {
-            None
-        };
-
-        // Spawn runtime for router->local indexer comm if requested
-        let _local_indexer_query_handle = local_indexer.as_ref().map(|local_indexer_ref| {
-            let component = component.clone();
-            let local_indexer = local_indexer_ref.clone();
-
-            component
-                .drt()
-                .runtime()
-                .secondary()
-                .spawn(start_worker_kv_query_endpoint(
-                    component,
-                    worker_id,
-                    dp_rank,
-                    local_indexer,
-                ))
-        });
-
-        let cancellation_token_clone = cancellation_token.clone();
-        let local_indexer_clone = local_indexer.clone();
-
-        if enable_local_indexer {
-            // When local indexer is enabled, use the event plane directly.
-            // EventPublisher handles transport selection (ZMQ or NATS) based on environment.
-            // Durability is provided by the local indexer's event buffer.
-            tracing::info!("Using event plane for KV event publishing (local_indexer mode)");
-            let component_clone = component.clone();
-            component.drt().runtime().secondary().spawn(async move {
-                let event_publisher =
-                    match EventPublisher::for_component(&component_clone, KV_EVENT_SUBJECT).await {
-                        Ok(publisher) => publisher,
-                        Err(e) => {
-                            tracing::error!("Failed to create event publisher: {}", e);
-                            return;
-                        }
-                    };
-
-                start_event_processor(
-                    EventPlanePublisher(event_publisher),
-                    worker_id,
-                    cancellation_token_clone,
-                    rx,
-                    local_indexer_clone,
-                    batching_timeout_ms,
-                )
-                .await
-            });
-        } else {
-            // When local indexer is disabled, use JetStream (NatsQueue) for durability.
-            let stream_name = create_kv_stream_name(&component, KV_EVENT_SUBJECT);
-            let nats_server = std::env::var(env_nats::NATS_SERVER)
-                .unwrap_or_else(|_| "nats://localhost:4222".to_string());
-            let mut nats_queue = NatsQueue::new_without_consumer(
-                stream_name,
-                nats_server,
-                std::time::Duration::from_secs(60), // 1 minute timeout
-            );
-
-            component.drt().runtime().secondary().spawn(async move {
-                if let Err(e) = nats_queue.connect().await {
-                    tracing::error!("Failed to connect NatsQueue: {e}");
-                    return;
-                }
-                start_event_processor_jetstream(
-                    JetStreamPublisher(nats_queue),
-                    worker_id,
-                    cancellation_token_clone,
-                    rx,
-                    local_indexer_clone,
-                    batching_timeout_ms,
-                )
-                .await
-            });
-        }
-
-        Ok(Self {
-            kv_block_size,
-            source,
-            cancellation_token,
-            worker_id,
-            tx,
-            next_event_id,
-        })
-    }
-
-    pub fn publish(&self, event: KvCacheEvent) -> Result<(), mpsc::error::SendError<KvCacheEvent>> {
-        let placement_event = PlacementEvent::local_gpu(self.worker_id, event);
-        match self.tx.send(placement_event) {
-            Ok(()) => Ok(()),
-            Err(err) => Err(mpsc::error::SendError(err.0.event)),
-        }
-    }
-
-    /// Get and increment the next event ID atomically.
-    /// Use this to assign monotonically increasing event IDs to events before publishing.
-    pub fn next_event_id(&self) -> u64 {
-        self.next_event_id.fetch_add(1, Ordering::SeqCst)
-    }
-
-    pub fn kv_block_size(&self) -> u32 {
-        self.kv_block_size
-    }
-
-    pub fn shutdown(&mut self) {
-        if !self.cancellation_token.is_cancelled() {
-            self.cancellation_token.cancel();
-        }
-
-        if let Some(source) = self.source.take() {
-            source.shutdown();
-        }
-    }
-}
-
-impl Drop for KvEventPublisher {
-    fn drop(&mut self) {
-        self.shutdown();
-    }
-}
-
+use super::*;
+#[allow(unused_imports)]
+use bytes::Bytes;
+#[allow(unused_imports)]
 use dynamo_kv_router::RouterEventSink;
-
-struct EventPlanePublisher(EventPublisher);
-
-impl RouterEventSink for EventPlanePublisher {
-    fn publish_event(&self, event: &RouterEvent) -> impl Future<Output = Result<()>> + Send {
-        self.0.publish(event)
-    }
-}
-
-struct JetStreamPublisher(NatsQueue);
-
-impl RouterEventSink for JetStreamPublisher {
-    fn publish_event(&self, event: &RouterEvent) -> impl Future<Output = Result<()>> + Send {
-        NatsQueue::publish_event(&self.0, KV_EVENT_SUBJECT, event)
-    }
-}
-
-/// Publishes a single [`KvCacheEvent`] to the event sink and, when present, the local indexer.
-/// Errors are logged and swallowed so the caller loop can continue uninterrupted.
-async fn emit<P: RouterEventSink>(
-    publisher: &P,
-    local_indexer: &Option<Arc<LocalKvIndexer>>,
-    worker_id: u64,
-    event: KvCacheEvent,
-) {
-    let router_event = RouterEvent::new(worker_id, event);
-    if let Some(indexer) = local_indexer
-        && let Err(e) = indexer.apply_event_with_buffer(router_event.clone()).await
-    {
-        tracing::warn!(worker_id, error = %e, "Failed to apply event to local indexer");
-    }
-    if let Err(e) = publisher.publish_event(&router_event).await {
-        tracing::error!(worker_id, error = %e, "Failed to publish event");
-    }
-}
-
-impl BatchingState {
-    /// Publishes any pending batch as a single [`RouterEvent`] and advances the monotonic
-    /// batch ID. No-ops when nothing is pending, so callers may call unconditionally.
-    async fn flush<P: RouterEventSink + Send + Sync + 'static>(
-        &mut self,
-        publisher: &P,
-        local_indexer: &Option<Arc<LocalKvIndexer>>,
-        worker_id: u64,
-    ) {
-        if !self.has_pending() {
-            return;
-        }
-        let id = self.next_publish_id;
-        let dp_rank = self.last_dp_rank;
-        if let Some(data) = self.pending_removed.take() {
-            emit(
-                publisher,
-                local_indexer,
-                worker_id,
-                KvCacheEvent {
-                    event_id: id,
-                    data: KvCacheEventData::Removed(data),
-                    dp_rank,
-                },
-            )
-            .await;
-        }
-        if let Some(data) = self.pending_stored.take() {
-            emit(
-                publisher,
-                local_indexer,
-                worker_id,
-                KvCacheEvent {
-                    event_id: id,
-                    data: KvCacheEventData::Stored(data),
-                    dp_rank,
-                },
-            )
-            .await;
-        }
-        // Consecutive batch IDs (1, 2, 3, …) keep downstream gap-detection happy.
-        self.next_publish_id += 1;
-        // Record when we flushed for stale-data detection on next event.
-        self.record_flush_time();
-    }
-}
-
-/// Batching loop: accumulates Removed/Stored events and flushes them as a single
-/// [`RouterEvent`] when any of the following conditions are met:
-/// - Event type switches (Removed ↔ Stored)
-/// - `dp_rank` changes between consecutive events
-/// - A `Stored` event's `parent_hash` breaks the sequential chain
-/// - The batch window expires (`Some(timeout_ms)`; `None` = disabled, flush every event)
-/// - Channel is closed or a cancellation signal is received
-async fn run_event_processor_loop<P: RouterEventSink + Send + Sync + 'static>(
-    publisher: P,
-    worker_id: u64,
-    cancellation_token: CancellationToken,
-    mut rx: mpsc::UnboundedReceiver<PlacementEvent>,
-    local_indexer: Option<Arc<LocalKvIndexer>>,
-    timeout_ms: Option<u64>,
-    max_batch_blocks: usize,
-) {
-    let mut batching_state = BatchingState::new();
-    // Track last raw input event_id for gap detection (dropped events before batching).
-    // The raw event_id is a globally monotonic counter assigned by the ZMQ listener,
-    // so any gap here means events were silently dropped (e.g. send error on the channel).
-    let mut last_raw_input_id: Option<u64> = None;
-
-    loop {
-        tokio::select! {
-            _ = cancellation_token.cancelled() => {
-                tracing::info!("KV Event source received cancellation signal");
-                batching_state.flush(&publisher, &local_indexer, worker_id).await;
-                break;
-            }
-            event = rx.recv() => {
-                let Some(placement_event) = event else {
-                    tracing::debug!("Event processor channel closed.");
-                    batching_state.flush(&publisher, &local_indexer, worker_id).await;
-                    break;
-                };
-
-                // Warn if the raw input event_id is not consecutive — events were dropped
-                // (e.g. channel send error) before they reached the batching layer.
-                let raw_event_id = placement_event.event.event_id;
-                if let Some(last_id) = last_raw_input_id
-                    && raw_event_id > last_id + 1
-                {
-                    let gap = raw_event_id - last_id - 1;
-                    tracing::warn!(
-                        worker_id,
-                        last_raw_input_id = last_id,
-                        raw_event_id,
-                        gap,
-                        "Input event gap detected: raw events dropped before batching"
-                    );
-                    // Increment Prometheus counter for dropped events (if initialized)
-                    if let Some(metrics) = kv_publisher_metrics() {
-                        metrics.increment_engines_dropped_events(worker_id, gap);
-                    } else {
-                        tracing::warn!(
-                            worker_id,
-                            gap,
-                            "Failed to record dropped events metric: metrics not initialized"
-                        );
-                    }
-                }
-                last_raw_input_id = Some(raw_event_id);
-
-                if !placement_event.placement.is_local_gpu() {
-                    tracing::trace!(
-                        worker_id,
-                        ?placement_event.placement,
-                        event_id = placement_event.event.event_id,
-                        "Skipping non-local-GPU placement event"
-                    );
-                    continue;
-                }
-
-                let event = placement_event.event;
-                tracing::trace!("Event processor for worker_id {} processing event: {:?}", worker_id, event.data);
-
-                let dp_rank_changed = batching_state.has_pending()
-                    && event.dp_rank != batching_state.last_dp_rank;
-
-                match event.data {
-                    KvCacheEventData::Removed(data) => {
-                        if batching_state.pending_stored.is_some() || dp_rank_changed {
-                            batching_state.flush(&publisher, &local_indexer, worker_id).await;
-                        }
-                        match &mut batching_state.pending_removed {
-                            Some(pending) => pending.block_hashes.extend(data.block_hashes),
-                            None => {
-                                batching_state.pending_removed = Some(data);
-                            }
-                        }
-                    }
-                    KvCacheEventData::Stored(data) => {
-                        // Flush if: type switch, dp_rank change, or the chain is broken
-                        // (new event's parent_hash doesn't continue from the last stored block).
-                        let should_flush = dp_rank_changed
-                            || batching_state.pending_removed.is_some()
-                            || batching_state.pending_stored.as_ref().is_some_and(|p| {
-                                data.parent_hash != p.blocks.last().map(|b| b.block_hash)
-                            });
-                        if should_flush {
-                            batching_state.flush(&publisher, &local_indexer, worker_id).await;
-                        }
-                        match &mut batching_state.pending_stored {
-                            // Only extend blocks; parent_hash stays fixed from the first event.
-                            Some(pending) => pending.blocks.extend(data.blocks),
-                            None => {
-                                batching_state.pending_stored = Some(data);
-                            }
-                        }
-                    }
-                    KvCacheEventData::Cleared => {
-                        batching_state.flush(&publisher, &local_indexer, worker_id).await;
-                        emit(&publisher, &local_indexer, worker_id, KvCacheEvent {
-                            event_id: batching_state.next_publish_id,
-                            data: KvCacheEventData::Cleared,
-                            dp_rank: event.dp_rank,
-                        }).await;
-                        batching_state.next_publish_id += 1;
-                    }
-                }
-
-                // Track dp_rank after the match so in-flight flushes use the old value.
-                batching_state.last_dp_rank = event.dp_rank;
-
-                // Flush after every event when disabled (None), or when the window has elapsed,
-                // or when the batch exceeds the max block count.
-                // The sleep arm only arms when batching is enabled; this covers the disabled path.
-                if batching_state.has_pending()
-                    && (timeout_ms.is_none_or(|ms| batching_state.is_timeout_elapsed(ms))
-                        || batching_state.pending_block_count() > max_batch_blocks)
-                {
-                    batching_state.flush(&publisher, &local_indexer, worker_id).await;
-                }
-            }
-            // if has some pending and has timeout, and no new events come in, then flush when timeout elapsed to prevent stale events
-            _ = tokio::time::sleep(
-                timeout_ms.map(|ms| batching_state.remaining_timeout(ms)).unwrap_or(Duration::from_secs(3600))
-            ), if timeout_ms.is_some() && batching_state.has_pending() => {
-                batching_state.flush(&publisher, &local_indexer, worker_id).await;
-            }
-        }
-    }
-}
-
-/// Batched event processor for ephemeral transports (NATS Core / ZMQ).
-async fn start_event_processor<P: RouterEventSink + Send + Sync + 'static>(
-    publisher: P,
-    worker_id: u64,
-    cancellation_token: CancellationToken,
-    rx: mpsc::UnboundedReceiver<PlacementEvent>,
-    local_indexer: Option<Arc<LocalKvIndexer>>,
-    batching_timeout_ms: Option<u64>,
-) {
-    run_event_processor_loop(
-        publisher,
-        worker_id,
-        cancellation_token,
-        rx,
-        local_indexer,
-        batching_timeout_ms,
-        DEFAULT_MAX_BATCH_BLOCKS,
-    )
-    .await
-}
-
-/// Batched event processor using JetStream (durable).
-async fn start_event_processor_jetstream<P: RouterEventSink + Send + Sync + 'static>(
-    publisher: P,
-    worker_id: u64,
-    cancellation_token: CancellationToken,
-    rx: mpsc::UnboundedReceiver<PlacementEvent>,
-    local_indexer: Option<Arc<LocalKvIndexer>>,
-    batching_timeout_ms: Option<u64>,
-) {
-    run_event_processor_loop(
-        publisher,
-        worker_id,
-        cancellation_token,
-        rx,
-        local_indexer,
-        batching_timeout_ms,
-        DEFAULT_MAX_BATCH_BLOCKS,
-    )
-    .await
-}
-
-/// Calculate exponential backoff duration based on consecutive error count
-fn calculate_backoff_ms(consecutive_errors: u32) -> u64 {
-    std::cmp::min(
-        INITIAL_BACKOFF_MS * 2_u64.pow(consecutive_errors.min(MAX_BACKOFF_EXPONENT)),
-        MAX_BACKOFF_MS,
-    )
-}
-
-pub async fn start_zmq_listener(
-    zmq_endpoint: String,
-    zmq_topic: String,
-    worker_id: WorkerId,
-    tx: mpsc::UnboundedSender<PlacementEvent>,
-    cancellation_token: CancellationToken,
-    kv_block_size: u32,
-    next_event_id: Arc<AtomicU64>,
-) {
-    tracing::debug!(
-        "KVEventPublisher connecting to ZMQ endpoint {} (topic '{}')",
-        zmq_endpoint,
-        zmq_topic
-    );
-
-    let warning_count = Arc::new(AtomicU32::new(0));
-
-    let mut socket = SubSocket::new();
-
-    // Subscribe to the requested topic (empty string == all topics)
-    if let Err(e) = socket.subscribe(&zmq_topic).await {
-        tracing::error!("Failed to subscribe on ZMQ socket: {}", e);
-        return;
-    }
-
-    // Connect to the ZMQ endpoint. SGLang binds locally, Dynamo connects.
-    // In multi-node setups, each node runs dynamo.sglang alongside local SGLang ranks,
-    // so ZMQ connections are always local. NATS handles cross-node event distribution.
-    if let Err(e) = socket.connect(&zmq_endpoint).await {
-        tracing::error!("Failed to connect ZMQ SUB socket to {zmq_endpoint}: {e}");
-        return;
-    }
-
-    let mut consecutive_errors = 0u32;
-    #[expect(unused_assignments)]
-    let mut exit_reason = "unknown";
-    let mut messages_processed = 0u64;
-
-    'main: loop {
-        tokio::select! {
-            biased;
-
-            // Check for cancellation
-            _ = cancellation_token.cancelled() => {
-                tracing::debug!("ZMQ listener received cancellation signal");
-                exit_reason = "cancellation token cancelled";
-                break 'main;
-            }
-
-            // Receive message
-            msg_result = socket.recv() => {
-                let Ok(msg) = msg_result else {
-                    let e = msg_result.unwrap_err();
-                    consecutive_errors += 1;
-
-                    if consecutive_errors >= MAX_CONSECUTIVE_ERRORS {
-                        tracing::error!(
-                            error=%e,
-                            consecutive_errors=%consecutive_errors,
-                            "Too many consecutive ZMQ errors, terminating listener"
-                        );
-                        exit_reason = "too many consecutive errors";
-                        break 'main;
-                    }
-
-                    // Simple exponential backoff with max exponent to prevent overflow
-                    let backoff_ms = calculate_backoff_ms(consecutive_errors);
-
-                    tracing::warn!(
-                        error=%e,
-                        consecutive_errors=%consecutive_errors,
-                        backoff_ms=%backoff_ms,
-                        "Error reading from ZMQ socket, applying exponential backoff"
-                    );
-
-                    tokio::time::sleep(Duration::from_millis(backoff_ms)).await;
-                    continue;
-                };
-                // Reset error count on successful message
-                consecutive_errors = 0;
-
-                // We expect multipart frames: [topic, seq, payload]
-                let mut frames: Vec<Vec<u8>> = msg.into_vec().into_iter().map(|frame| frame.to_vec()).collect();
-
-                if frames.len() != 3 {
-                    tracing::warn!("Received unexpected ZMQ frame count: expected 3, actual {}", frames.len());
-                    continue;
-                }
-
-                // Extract the payload and sequence number.
-                let payload = frames.pop().unwrap();
-                let seq_bytes = frames.pop().unwrap();
-
-                if seq_bytes.len() != 8 {
-                    tracing::warn!("Invalid sequence number byte length: expected 8, actual {}", seq_bytes.len());
-                    continue;
-                }
-
-                // Note: We extract the engine's sequence number for logging but use our own
-                // internal monotonic counter for event_id to ensure per-dp_rank monotonicity
-                let engine_seq = u64::from_be_bytes(seq_bytes.try_into().unwrap());
-
-                // Decode our batch of events.
-                let batch_result = rmps::from_slice::<KvEventBatch>(&payload);
-                let Ok(batch) = batch_result else {
-                    let e = batch_result.unwrap_err();
-                    tracing::warn!("Failed to decode KVEventBatch msgpack: {e}");
-                    continue;
-                };
-
-                tracing::trace!(
-                    "ZMQ listener on {} received batch with {} events (engine_seq={}, dp_rank={})",
-                    zmq_endpoint,
-                    batch.events.len(),
-                    engine_seq,
-                    batch.data_parallel_rank.unwrap_or(0)
-                );
-
-                let dp_rank = batch.data_parallel_rank.unwrap_or(0).cast_unsigned();
-                for raw_event in batch.events.into_iter() {
-                    // Use shared monotonic event_id counter instead of engine's sequence number
-                    let event_id = next_event_id.fetch_add(1, Ordering::SeqCst);
-                    let worker = WorkerWithDpRank::new(worker_id, dp_rank);
-                    let event = convert_event(
-                        raw_event,
-                        event_id,
-                        kv_block_size,
-                        worker,
-                        &warning_count,
-                    );
-                    if tx.send(event).is_err() {
-                        tracing::warn!("Failed to send message to channel - receiver dropped");
-                        exit_reason = "channel receiver dropped";
-                        break 'main;
-                    }
-                    messages_processed += 1;
-                }
-            }
-        }
-    }
-    tracing::debug!(
-        "ZMQ listener exiting, reason: {}, messages processed: {}",
-        exit_reason,
-        messages_processed
-    );
-}
-
-// -------------------------------------------------------------------------
-// Metrics Publishers ------------------------------------------------------
-// -------------------------------------------------------------------------
-
-/// Metrics data passed through the channel for NATS publishing
-#[derive(Debug, Clone, Default, PartialEq)]
-struct WorkerMetrics {
-    dp_rank: DpRank,
-    active_decode_blocks: u64,
-}
-
-pub struct WorkerMetricsPublisher {
-    tx: tokio::sync::watch::Sender<WorkerMetrics>,
-    rx: tokio::sync::watch::Receiver<WorkerMetrics>,
-}
-
-impl WorkerMetricsPublisher {
-    pub fn new() -> Result<Self> {
-        let (tx, rx) = tokio::sync::watch::channel(WorkerMetrics::default());
-        Ok(WorkerMetricsPublisher { tx, rx })
-    }
-
-    /// Publish worker metrics for load monitoring.
-    ///
-    /// # Arguments
-    /// * `dp_rank` - Data parallel rank of the worker (None defaults to 0)
-    /// * `active_decode_blocks` - Number of active KV cache blocks
-    pub fn publish(&self, dp_rank: Option<DpRank>, active_decode_blocks: u64) -> Result<()> {
-        let metrics = WorkerMetrics {
-            dp_rank: dp_rank.unwrap_or(0),
-            active_decode_blocks,
-        };
-        tracing::trace!(
-            "Publish metrics: dp_rank={}, active_decode_blocks={}",
-            metrics.dp_rank,
-            metrics.active_decode_blocks
-        );
-        self.tx
-            .send(metrics)
-            .map_err(|_| anyhow::anyhow!("metrics channel closed"))
-    }
-
-    pub async fn create_endpoint(&self, component: Component) -> Result<()> {
-        let worker_id = component.drt().connection_id();
-        self.start_nats_metrics_publishing(component.namespace().clone(), worker_id);
-        Ok(())
-    }
-
-    /// Starts a background task to publish metrics over NATS
-    ///
-    /// This task monitors metric changes (specifically active_decode_blocks)
-    /// and publishes stable metrics to NATS after they've been unchanged for 1ms.
-    fn start_nats_metrics_publishing(&self, namespace: Namespace, worker_id: u64) {
-        let nats_rx = self.rx.clone();
-
-        tokio::spawn(async move {
-            let event_publisher =
-                match EventPublisher::for_namespace(&namespace, KV_METRICS_SUBJECT).await {
-                    Ok(publisher) => publisher,
-                    Err(e) => {
-                        tracing::error!("Failed to create metrics publisher: {}", e);
-                        return;
-                    }
-                };
-
-            let mut rx = nats_rx;
-            let mut last_metrics: Option<WorkerMetrics> = None;
-            let mut pending_publish: Option<WorkerMetrics> = None;
-            let mut publish_timer =
-                Box::pin(tokio::time::sleep(tokio::time::Duration::from_secs(0)));
-            publish_timer.as_mut().reset(tokio::time::Instant::now()); // Complete immediately
-
-            loop {
-                tokio::select! {
-                    // Handle metrics changes
-                    result = rx.changed() => {
-                        if result.is_err() {
-                            tracing::debug!(
-                                "Metrics publisher sender dropped, stopping NATS background task"
-                            );
-                            break;
-                        }
-
-                        let metrics = rx.borrow_and_update().clone();
-
-                        // Check if metrics have changed
-                        let has_changed = last_metrics.as_ref() != Some(&metrics);
-
-                        // If metrics changed, schedule a publish
-                        if has_changed {
-                            pending_publish = Some(metrics.clone());
-                            last_metrics = Some(metrics);
-
-                            // Start the 1ms timer
-                            publish_timer.as_mut().reset(
-                                tokio::time::Instant::now() + tokio::time::Duration::from_millis(1)
-                            );
-                        }
-                    }
-                    // Timer expired - publish if we have pending metrics
-                    _ = &mut publish_timer => {
-                        if let Some(metrics) = pending_publish.take() {
-                            let active_load = ActiveLoad {
-                                worker_id,
-                                dp_rank: metrics.dp_rank,
-                                active_decode_blocks: Some(metrics.active_decode_blocks),
-                                active_prefill_tokens: None,
-                            };
-
-                            if let Err(e) = event_publisher.publish(&active_load).await {
-                                tracing::warn!("Failed to publish metrics: {}", e);
-                            }
-                        }
-
-                        // Reset timer to pending state to avoid tight loop
-                        // It will be reset to 1ms when metrics actually change
-                        publish_timer.as_mut().reset(
-                            tokio::time::Instant::now() + tokio::time::Duration::from_secs(3600)
-                        );
-                    }
-                }
-            }
-        });
-    }
-}
-
-// -------------------------------------------------------------------------
-// Testing -----------------------------------------------------------------
-// -------------------------------------------------------------------------
+#[allow(unused_imports)]
+use rmp_serde as rmps;
+#[allow(unused_imports)]
+use std::future::Future;
+#[allow(unused_imports)]
+use std::sync::atomic::{AtomicU32, AtomicU64, Ordering};
+use std::time::Duration;
+#[allow(unused_imports)]
+use zeromq::{PubSocket, Socket, SocketSend, ZmqMessage};
 
 #[cfg(test)]
 mod test_event_processing {
@@ -1459,9 +427,8 @@ mod test_event_processing {
 #[cfg(test)]
 mod tests_startup_helpers {
     use super::*;
-    use crate::kv_router::KvIndexer;
     use bytes::Bytes;
-    use dynamo_kv_router::indexer::{GetWorkersRequest, KvIndexerInterface};
+    use dynamo_kv_router::indexer::{GetWorkersRequest, KvIndexer, KvIndexerInterface};
     use dynamo_kv_router::protocols::{ExternalSequenceBlockHash, LocalBlockHash};
     use std::sync::{Arc, Mutex};
     use zeromq::{PubSocket, Socket, SocketSend, ZmqMessage};
